databricks/app/requirements.txt
streamlit>=1.35.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
numpy>=1.26.0
